Rename test form helpers to kebab-case
