+++ test/controllers/company-ukother.controller.test.js
@@ -4,7 +4,7 @@
 const proxyquire = require('proxyquire')
 const { render } = require('../nunjucks')
 const { hqLabels } = require('../../src/labels/company-labels')
-const { expectTextFieldWithLabel, expectDropdownWithLabel, expectHiddenField, expectRadioWithLabel, expectTextAreaWithLabel } = require('../formhelpers')
+const { expectTextFieldWithLabel, expectDropdownWithLabel, expectHiddenField, expectRadioWithLabel, expectTextAreaWithLabel } = require('../form-helpers')
 const next = function (error) {
   throw Error(error)
 }
